Add get 'root' route to allow GET request on ..
index page. In static_pages #index use flickr gem to access the
environment variables FLICKR_API_KEY and FLICKR_SHARED_SECRET stored
safely in application.yml using figaro gem.
Flickr gem to populate @url_array with the urls of photos. In the
index view, img elements with the urls as their src attributes are to
display the photos. Rescue any flickr errors with a flash message in the controller.